# tools/openapi.yaml
openapi: 3.0.3
info:
  version: 0.9.6
  title: LANraragi API
...
paths:
...
  /tempfolder:
    delete:
      security:
        - api_key: []
      operationId: cleanTempfolder
      x-mojo-to: api-other#clean_tempfolder
      summary: Clean the Temporary Folder
      description: Cleans the server's temporary folder.
      tags:
        - misc
      responses:
        '200':
          description: Cleanup result
          content:
            application/json:
              schema:
                type: object
                properties:
                  operation:
                    type: string
                    enum: 
                      - cleantemp
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
                  error:
                    type: string
                    nullable: true
                    description: Error message if something went wrong during cleanup. (The call will still return 200)
                  newsize:
                    type: integer
                    description: Current size of the temporary folder post-cleanup.
                example:
                  operation: cleantemp
                  error: null
                  newsize: 0
                  success: 1
  /download_url:
    post:
      security:
        - api_key: []
      operationId: downloadUrl
      x-mojo-to: api-other#download_url
      summary: Queue a URL download
      description: Add a URL to be downloaded by the server and added to its library.
      tags:
        - misc
      parameters:
        - name: url
          in: query
          description: URL to download
          required: false
          schema:
            type: string
            format: uri
        - name: catid
          in: query
          required: false
          description: Category ID to add the downloaded URL to.
          schema:
            type: string
            minLength: 14
            maxLength: 14
      requestBody:
        required: false
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  description: URL to download
                  type: string
                  format: uri
                catid:
                  type: string
                  description: Category ID to add the downloaded URL to.
                  minLength: 14
                  maxLength: 14
      responses:
        '200':
          description: Enqueued job
          content:
            application/json:
              schema:
                type: object
                properties:
                  operation:
                    type: string
                    enum: 
                      - download_url
                  url:
                    type: string
                    description: URL queued for download
                  category:
                    type: string
                    nullable: true
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
                  job:
                    type: integer
                example:
                  job: 86
                  operation: download_url
                  success: 1
                  url: https://example.com
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
  /regen_thumbs:
    post:
      security:
        - api_key: []
      operationId: regenThumbnails
      x-mojo-to: api-other#regen_thumbnails
      summary: Regenerate Thumbnails
      description: Queue a Minion job to regenerate missing/all thumbnails on the server.
      tags:
        - misc
      parameters:
        - name: force
          in: query
          description: Whether to generate all thumbnails, or only the missing ones.
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Enqueued job
          content:
            application/json:
              schema:
                type: object
                properties:
                  operation:
                    type: string
                    enum: 
                      - regen_thumbnails
                  success:
                    type: integer
                    enum:
                      - 0
                      - 1
                  job:
                    type: integer
                example:
                  job: 3
                  operation: regen_thumbnails
                  success: 1
...
  /minion/{jobid}:
    get:
      operationId: minionJobStatus
      x-mojo-to: api-minion#minion_job_status
      summary: Get the basic status of a Minion Job
      description: >- 
        For a given Minion job ID, check whether it succeeded or failed.  

        Minion jobs are ran for various occasions like thumbnails, cache warmup and handling incoming files.  

        For some jobs, you can check the notes field for progress information. Look at https://docs.mojolicious.org/Minion/Guide#Job-progress for more information.
      tags:
        - minion
      parameters:
        - name: jobid
          in: path
          required: true
          description: ID of the Job to query status for.
          schema:
            type: integer
      responses:
        '200':
          description: Job status data
          content:
            application/json:
              schema:
                type: object
                properties:
                  task:
                    type: string
                  state:
                    type: string
                    enum:
                      - inactive
                      - active
                      - finished
                      - failed
                  notes:
                    type: object
                    nullable: true
                    description: Arbitrary data attached to the Job.
                  result:
                    type: object
                    nullable: true
                    description: Finish payload returned by the task on successful completion.
                  error:
                    type: string
                    nullable: true
                example:
                  state: finished
                  task: handle_upload
                  notes: {}
                  result:
                    example_key: This is the task finish payload.
                  error: null
  /minion/{jobid}/detail:
    get:
      security:
        - api_key: []
      operationId: minionJobDetail
      x-mojo-to: api-minion#minion_job_detail
      summary: 🔑 Get the full status of a Minion Job
      description: >- 
        Get the status of a Minion Job.  

        This API is there for internal usage mostly, but you can use it to get detailed status for jobs like plugin runs or URL downloads.
      tags:
        - minion
      parameters:
        - name: jobid
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Job detail
          content:
            application/json:
              schema:
                type: object
                example:
                    args: [/tmp/QF3UCnKdMr/myfile.zip]
                    attempts: 1
                    children: []
                    created: 1601145004
                    delayed: 1601145004
                    expires: null
                    finished: 1601145004
                    id: 7
                    lax: 0
                    notes: {}
                    parents: []
                    priority: 0
                    queue: default
                    result: 
                      id: 75d18ce470dc99f83dc355bdad66319d1f33c82b
                      message: This file already exists in the Library.
                      success: 0
                    retried: null
                    retries: 0
                    started: 1601145004
                    state: finished
                    task: handle_upload
                    time: 1601145005
                    worker: 1
  /minion/{jobname}/queue:
    post:
      security:
        - api_key: []
      operationId: queueMinionJob
      x-mojo-to: api-minion#queue_minion_job
      summary: 🔑 Queue a Minion job
      description: >- 
        Queue a Job with the specified type and parameters.  

        See LANraragi::Utils::Minion for all the available types of jobs and the parameters they require.  


        There's no API contract in place for whether a Job type exists on a given server version, so I wouldn't recommend using this unless you have a good reason to.  
      tags:
        - minion
      parameters:
        - name: jobname
          in: path
          required: true
          description: Type of the job to instantiate.
          schema:
            type: string
            enum:
              - thumbnail_task
              - tank_thumbnail_task
              - page_thumbnails
              - regen_all_thumbnails
              - find_duplicates
              - build_stat_hashes
              - handle_upload
              - download_url
              - run_plugin
        - name: args
          in: query
          required: true
          description: Arguments as a JSON array string
          schema:
            type: string
        - name: priority
          in: query
          required: false
          description: Priority of the Minion job. The higher the number, the more important the job is.
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Enqueued job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MinionJobResponse'
...
components:
  securitySchemes:
    api_key:
      type: apiKey
      name: Authorization
      in: header
      description: 'Use Authorization: Bearer <base64(api_key)>'
  schemas:
...
    OperationResponse:
      type: object
      required:
        - operation
        - success
      properties:
        operation:
          type: string
          description: Name of operation
        error:
          type: string
          description: Error message if any
        successMessage:
          type: string
          description: Success message if any
        success:
          type: integer
          description: Returns 1 if operation was successful, else 0
          enum:
            - 0
            - 1
    MinionJobResponse:
      type: object
      description: JSON object for a queued Minion job.
      required:
        - job
        - operation
        - success
      properties:
        job:
          type: integer
          description: Minion job ID
        operation:
          type: string
          description: Name of operation
        success:
          type: integer
          description: Returns 1 if operation was successful, else 0
          enum: 
            - 1
            - 0
...
servers:
  - url: https://lrr.tvc-16.science/api